requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg[binary]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
email-validator

